+++ GANDLF/utils/__init__.py
@@ -15,7 +15,7 @@ from .torch import (
     get_class_imbalance_weights,
 )
 
-from .io import (
+from .write_parse import (
     writeTrainingCSV,
     parseTrainingCSV,
 )
